adding some styling and date manipulations
