+++ trending-repos/src/components/Repository.js
@@ -1,9 +1,10 @@
 import React from 'react';
 
 import '../styles/Repository.css';
+import { numberOfDaysSinceRepoCreation } from '../utils/dateManipulations';
 
 function Repository({repo}){
-    const {name, owner, description, stargazers_count, open_issues_count} = repo;
+    const {name, owner, description, stargazers_count, open_issues_count, created_at} = repo;
     return (
       <div className="repo">
           <div className="owner-avatar">
@@ -13,9 +14,9 @@ function Repository({repo}){
             <h1>{name}</h1>
             <p>{description}</p>
             <div className="repo-stats">
-                <p>Stars: {stargazers_count}</p>
-                <p>Issues: {open_issues_count}</p>
-                <p>Submitted 30 days ago by tensorflow</p>
+                <div>Stars: {stargazers_count}</div>
+                <div>Issues: {open_issues_count}</div>
+                <p>Submitted {numberOfDaysSinceRepoCreation(created_at)} days ago by {owner.login}</p>
             </div>
           </div>
       </div>
