+++ trending-repos/src/styles/Repository.css
@@ -24,8 +24,15 @@ img {
 .repo-stats{
     display: flex;
     flex-direction: row;
+    align-items: center;
 }
 
-.repo-stats p {
+.repo-stats div {
+    height: 20px;
     margin-right: 1vw;
+    border: 2px solid black;
+    text-align: center;
+    vertical-align: middle;
+    line-height: 20px;
+    padding: 3px;
 }
